split API up by type (second method)
