+++ routes/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+routes = Blueprint("routes", __name__)
+
+from .fire import *
+from .permafrost import *
